move workers to own module
